Move isDomSplitDiff() to dom_functions
https://github.com/sourcegraph/sourcegraph/pull/3494#discussion_r277717420
