+++ client/browser/src/libs/github/util.test.ts
@@ -1,6 +1,4 @@
-import { startCase } from 'lodash'
-import { readFile } from 'mz/fs'
-import { isDomSplitDiff, parseGitHubHash } from './util'
+import { parseGitHubHash } from './util'
 
 describe('util', () => {
     describe('parseGitHubHash()', () => {
@@ -11,59 +9,5 @@ describe('util', () => {
         it('handles invalid value', () => expect(parseGitHubHash('#Lfoo')).toBe(undefined))
         it('allows extra after', () =>
             expect(parseGitHubHash('#L123-L456-foo')).toEqual({ startLine: 123, endLine: 456 }))
-    })
-
-    describe('isDomSplitDiff()', () => {
-        for (const version of ['github.com', 'ghe-2.14.11']) {
-            describe(`Version ${version}`, () => {
-                const views = [
-                    {
-                        view: 'pull-request',
-                        url: 'https://github.com/sourcegraph/sourcegraph/pull/2672/files',
-                        hasSplitView: true,
-                    },
-                    {
-                        view: 'commit',
-                        url:
-                            'https://github.com/sourcegraph/sourcegraph/commit/2c74f329fd03008fa0b446cd5e53234715dae3dc',
-                        hasSplitView: true,
-                    },
-                    {
-                        view: 'pull-request-discussion',
-                        url: 'https://github.com/sourcegraph/sourcegraph/pull/2672/',
-                        hasSplitView: false,
-                    },
-                ]
-                for (const { view, url, hasSplitView } of views) {
-                    describe(`${startCase(view)} page`, () => {
-                        beforeEach(() => {
-                            jsdom.reconfigure({ url })
-                        })
-                        for (const extension of ['vanilla', 'refined-github']) {
-                            describe(startCase(extension), () => {
-                                if (hasSplitView) {
-                                    it('should return true for split view', async () => {
-                                        document.body.innerHTML = await readFile(
-                                            `${__dirname}/__fixtures__/${version}/${view}/${extension}/split/page.html`,
-                                            'utf-8'
-                                        )
-                                        expect(isDomSplitDiff(document.querySelector('.file') as HTMLElement)).toBe(
-                                            true
-                                        )
-                                    })
-                                }
-                                it('should return false for unified view', async () => {
-                                    document.body.innerHTML = await readFile(
-                                        `${__dirname}/__fixtures__/${version}/${view}/${extension}/unified/page.html`,
-                                        'utf-8'
-                                    )
-                                    expect(isDomSplitDiff(document.querySelector('.file') as HTMLElement)).toBe(false)
-                                })
-                            })
-                        }
-                    })
-                }
-            })
-        }
     })
 })
